agents/amazon-purchase-agent/requirements.txt
anthropic>=0.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
playwright>=1.44.0

# KIT Onboarding Agent
requests>=2.31.0
PyYAML>=6.0
PyMuPDF>=1.24.0      # PDF → image conversion (fitz); free & open-source
